+++ internal/domain/order/pg_query.go
@@ -24,7 +24,7 @@ func (s *PGStore) List(ctx context.Context, q OrderQuery) ([]OrderListItem, erro
 		LEFT JOIN user_addresses ua ON o.address_id = ua.id
 		WHERE ($1 = '' OR o.order_no ILIKE '%' || $1 || '%')
 		  AND ($2 = '' OR o.status = $2)
-		  AND ($3 = 0 OR o.customer_id = $3)
+		  AND ($3::bigint = 0 OR o.customer_id = $3)
 		ORDER BY o.id DESC
 		LIMIT $4 OFFSET $5`,
 		q.Keyword, q.Status, q.CustomerID, limit, q.Offset)
